+++ music_assistant/controllers/music/tracks.py
@@ -11,8 +11,8 @@ from music_assistant.helpers.compare import (
     compare_track,
 )
 from music_assistant.helpers.util import create_sort_name, merge_dict, merge_list
-from music_assistant.helpers.web import json_serializer
-from music_assistant.music.models import (
+from music_assistant.helpers.json import json_serializer
+from music_assistant.models.media_items import (
     ItemMapping,
     MediaControllerBase,
     MediaType,
