+++ music_assistant/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Package with Music Providers."""
