+++ music_assistant/providers/filesystem.py
@@ -6,9 +6,9 @@ from typing import List, Optional
 
 import taglib
 
-from music_assistant.helpers.errors import InvalidDataError
+from music_assistant.models.errors import InvalidDataError
 from music_assistant.helpers.util import parse_title_and_version
-from music_assistant.music.models import (
+from music_assistant.models.media_items import (
     Album,
     Artist,
     ContentType,
@@ -21,7 +21,7 @@ from music_assistant.music.models import (
     StreamType,
     Track,
 )
-from music_assistant.music_providers.model import MusicProvider
+from music_assistant.models.provider import MusicProvider
 
 
 class FileProvider(MusicProvider):
