+++ music_assistant/providers/qobuz.py
@@ -11,9 +11,9 @@ from music_assistant.constants import EventType
 from music_assistant.helpers.app_vars import (  # pylint: disable=no-name-in-module
     get_app_var,
 )
-from music_assistant.helpers.errors import LoginFailed
+from music_assistant.models.errors import LoginFailed
 from music_assistant.helpers.util import parse_title_and_version, try_parse_int
-from music_assistant.music.models import (
+from music_assistant.models.media_items import (
     Album,
     AlbumType,
     Artist,
@@ -27,7 +27,7 @@ from music_assistant.music.models import (
     StreamType,
     Track,
 )
-from music_assistant.music_providers.model import MusicProvider
+from music_assistant.models.provider import MusicProvider
 
 
 class QobuzProvider(MusicProvider):
