+++ music_assistant/providers/spotify/spotty/windows/spotty.exe
